Updated CreditClass, DebitCard, PaymentStatus class and Created ProcessPayment and UseCase class
