+++ src/main/java/app/pedidos/WebApp/domain/models/PaymentStatus.java
@@ -1,0 +1,10 @@
+package app.pedidos.WebApp.domain.models;
+
+import lombok.Data;
+
+@Data
+public class PaymentStatus {
+
+    private String status;
+    private String description;
+}
