--- src/main/java/models/PaymentStatus.java
@@ -1,15 +1,0 @@
-package models;
-
-public class PaymentStatus {
-
-    private String status;
-    private String description;
-
-    public String getStatus() {
-        return status;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-}
