--- src/main/java/models/PaymentType.java
@@ -1,12 +1,0 @@
-package models;
-
-import java.math.BigDecimal;
-
-public interface PaymentType {
-
-   public String payment(BigDecimal statusName);
-
-   public PaymentStatus checkStatus(String hash);
-
-
-}
